requirements.txt
Django==5.0.3
djangorestframework==3.14.0
fitz==0.0.1.dev2
protobuf==6.30.1
python-decouple==3.8
